Implement post_save and post_delete
