+++ elyra/pipeline/processor.py
@@ -80,10 +80,6 @@ class PipelineProcessorRegistry(SingletonConfigurable):
     def is_valid_processor(self, processor_type: str) -> bool:
         return processor_type in self._processors.keys()
 
-    @property
-    def processor_instances(self):
-        return list(self._processors.values())
-
 
 class PipelineProcessorManager(SingletonConfigurable):
     _registry: PipelineProcessorRegistry
@@ -99,11 +95,6 @@ class PipelineProcessorManager(SingletonConfigurable):
 
     def is_supported_runtime(self, processor_type: str) -> bool:
         return self._registry.is_valid_processor(processor_type)
-
-    def update_component_cache(self):
-        for processor in self._registry.processor_instances:
-            if isinstance(processor.component_registry, CachedComponentRegistry):
-                processor.component_registry.update_cache()
 
     async def get_components(self, processor_type):
         processor = self._get_processor_for_runtime(processor_type)
